+++ misc/dbt-materialize/dbt/include/materialize/macros/deploy/deploy_promote.sql
@@ -13,7 +13,7 @@
 -- See the License for the specific language governing permissions and
 -- limitations under the License.
 
-{% macro deploy_promote(wait=False, poll_interval=15) %}
+{% macro deploy_promote(wait=False, poll_interval=15, dry_run=False) %}
 {#
   Performs atomic deployment of current dbt targets to production,
   based on the deployment configuration specified in the dbt_project.yml file.
@@ -23,10 +23,12 @@
   to maintain consistency and prevent partial updates.
 
   ## Arguments
-  - `wait` (boolean, optional): Waits for the deployment to be fully hyrated.
+  - `wait` (boolean, optional): Waits for the deployment to be fully hydrated.
       Defaults to false. It is recommended you call `deploy_await` manually and
       run additional validation checks before promoting a deployment to production.
   - `poll_interval` (integer): The interval, in seconds, between each readiness check.
+  - `dry_run` (boolean, optional): When True, prints out the commands that would
+    be executed as part of the deployment workflow (without executing them).
 
   ## Returns
   None: This macro performs deployment actions but does not return a value.
@@ -56,14 +58,6 @@
     {% if not schema_exists(deploy_schema) %}
         {{ exceptions.raise_compiler_error("Deployment schema " ~ deploy_schema ~ " does not exist") }}
     {% endif %}
-    {% if schema_contains_sinks(deploy_schema) %}
-        {{ exceptions.raise_compiler_error("""
-        Deployment schema " ~ deploy_schema ~ " contains sinks. This is not currently
-        supported by the deploy_promote macro.
-
-        If this feature is important to you, please reach out!
-        """) }}
-    {% endif %}
 {% endfor %}
 
 {% for cluster in clusters %}
@@ -74,37 +68,49 @@
     {% if not cluster_exists(deploy_cluster) %}
         {{ exceptions.raise_compiler_error("Deployment cluster " ~ deploy_cluster ~ " does not exist") }}
     {% endif %}
-    {% if cluster_contains_sinks(deploy_cluster) %}
-        {{ exceptions.raise_compiler_error("""
-        Deployment cluster " ~ deploy_cluster ~ " contains sinks. This is not currently
-        supported by the deploy_promote macro.
-
-        If this feature is important to you, please reach out!
-        """) }}
-    {% endif %}
 {% endfor %}
 
 {% if wait %}
     {{ deploy_await(poll_interval) }}
 {% endif %}
 
-{% call statement('swap', fetch_result=True, auto_begin=False) -%}
-BEGIN;
+{% if not dry_run %}
+    {% call statement('swap', fetch_result=True, auto_begin=False) -%}
+    BEGIN;
 
-{% for schema in schemas %}
-    {% set deploy_schema = schema ~ "_dbt_deploy" %}
-    {{ log("Swapping schemas " ~ schema ~ " and " ~ deploy_schema, info=True) }}
-    ALTER SCHEMA {{ adapter.quote(schema) }} SWAP WITH {{ adapter.quote(deploy_schema) }};
-{% endfor %}
+    {% for schema in schemas %}
+        {% set deploy_schema = schema ~ "_dbt_deploy" %}
+        {{ log("Swapping schemas " ~ schema ~ " and " ~ deploy_schema, info=True) }}
+        ALTER SCHEMA {{ adapter.quote(schema) }} SWAP WITH {{ adapter.quote(deploy_schema) }};
+    {% endfor %}
 
-{% for cluster in clusters %}
-    {% set deploy_cluster = adapter.generate_final_cluster_name(cluster, force_deploy_suffix=True) %}
-    {{ log("Swapping clusters " ~ adapter.generate_final_cluster_name(cluster) ~ " and " ~ deploy_cluster, info=True) }}
-    ALTER CLUSTER {{ adapter.quote(cluster) }} SWAP WITH {{ adapter.quote(deploy_cluster) }};
-{% endfor %}
+    {% for cluster in clusters %}
+        {% set deploy_cluster = adapter.generate_final_cluster_name(cluster, force_deploy_suffix=True) %}
+        {{ log("Swapping clusters " ~ adapter.generate_final_cluster_name(cluster) ~ " and " ~ deploy_cluster, info=True) }}
+        ALTER CLUSTER {{ adapter.quote(cluster) }} SWAP WITH {{ adapter.quote(deploy_cluster) }};
+    {% endfor %}
 
-COMMIT;
-{%- endcall %}
+    COMMIT;
+    {%- endcall %}
+{% else %}
+    {{ log("Starting dry run...", info=True) }}
+    {% for schema in schemas %}
+        {% set deploy_schema = schema ~ "_dbt_deploy" %}
+        {{ log("DRY RUN: Swapping schemas " ~ schema ~ " and " ~ deploy_schema, info=True) }}
+        {{ log("DRY RUN: ALTER SCHEMA " ~ adapter.quote(schema) ~ " SWAP WITH " ~ adapter.quote(deploy_schema), info=True) }}
+    {% endfor %}
+
+    {% for cluster in clusters %}
+        {% set deploy_cluster = adapter.generate_final_cluster_name(cluster, force_deploy_suffix=True) %}
+        {{ log("DRY RUN: Swapping clusters " ~ adapter.generate_final_cluster_name(cluster) ~ " and " ~ deploy_cluster, info=True) }}
+        {{ log("DRY RUN: ALTER CLUSTER " ~ adapter.quote(cluster) ~ " SWAP WITH " ~ adapter.quote(deploy_cluster), info=True) }}
+    {% endfor %}
+    {{ log("Dry run completed. The statements above were **not** executed against Materialize.", info=True) }}
+{% endif %}
+
+{% set sinks_to_alter = discover_sinks(schemas) %}
+{{ process_sinks(sinks_to_alter, dry_run) }}
+
 {% endmacro %}
 
 {% macro cluster_contains_sinks(cluster) %}
@@ -134,5 +140,58 @@ COMMIT;
     {% set count = run_query(query) %}
     {% if execute %}
         {{ return(count.rows[0][0]) }}
+    {% endif %}
+{% endmacro %}
+
+{% macro get_current_database() %}
+    {% set current_database_query = "SELECT current_database()" %}
+    {% set results = run_query(current_database_query) %}
+    {% if execute %}
+        {% set current_database = results.rows[0][0] %}
+        {{ return(current_database) }}
+    {% else %}
+        {{ return(None) }}
+    {% endif %}
+{% endmacro %}
+
+{% macro discover_sinks(schemas) %}
+    {% set sinks_to_alter = [] %}
+    {% for node in graph.nodes.values() | selectattr("resource_type", "equalto", "model") | selectattr("config.materialized", "equalto", "sink") %}
+        {% set upstream_node = graph.nodes[node.depends_on.nodes[0]] %}
+        {% set upstream_schema = upstream_node.schema %}
+
+        {% if upstream_schema in schemas %}
+            {% set sink_database = node.database %}
+            {% set sink_schema = node.schema %}
+            {% set sink_name = node.name %}
+            {% set new_upstream_relation = adapter.quote(upstream_node.database) ~ '.' ~ adapter.quote(upstream_schema) ~ '.' ~ adapter.quote(upstream_node.name) %}
+            {% set sink = {
+                "database": sink_database,
+                "schema": sink_schema,
+                "name": sink_name,
+                "new_upstream_relation": new_upstream_relation
+            } %}
+            {% do sinks_to_alter.append(sink) %}
+        {% endif %}
+    {% endfor %}
+    {{ return(sinks_to_alter) }}
+{% endmacro %}
+
+{% macro process_sinks(sinks, dry_run) %}
+    {% if sinks|length > 0 %}
+        {% for sink in sinks %}
+            {% if sink['database'] and sink['schema'] and sink['name'] and sink['new_upstream_relation'] %}
+                {% if not dry_run %}
+                    {% call statement('alter_sink_' ~ loop.index, fetch_result=True, auto_begin=False) %}
+                        {{ log("Running ALTER SINK " ~ adapter.quote(sink['database']) ~ "." ~ adapter.quote(sink['schema']) ~ "." ~ adapter.quote(sink['name']) ~ " SET FROM " ~ sink['new_upstream_relation'], info=True) }}
+                        ALTER SINK {{ adapter.quote(sink['database']) }}.{{ adapter.quote(sink['schema']) }}.{{ adapter.quote(sink['name']) }} SET FROM {{ sink['new_upstream_relation'] }};
+                    {% endcall %}
+                {% else %}
+                    {{ log("DRY RUN: ALTER SINK " ~ adapter.quote(sink['database']) ~ "." ~ adapter.quote(sink['schema']) ~ "." ~ adapter.quote(sink['name']) ~ " SET FROM " ~ sink['new_upstream_relation'], info=True) }}
+                {% endif %}
+            {% endif %}
+        {% endfor %}
+    {% else %}
+        {{ log("No sinks to process.", info=True) }}
     {% endif %}
 {% endmacro %}