+++ misc/dbt-materialize/tests/adapter/fixtures.py
@@ -30,6 +30,12 @@ test_materialized_view_index = """
     SELECT * FROM (VALUES ('chicken', 'pig'), ('cow', 'horse')) _ (a, b)
 """
 
+test_materialized_view_deploy = """
+{{ config(materialized='materialized_view') }}
+
+    SELECT val + 1 AS val FROM source_table
+"""
+
 test_view = """
 {{ config(materialized='view') }}
 
@@ -102,7 +108,8 @@ FOR ALL TABLES;
 
 test_sink = """
 {{ config(
-    materialized='sink'
+    materialized='sink',
+    pre_hook="CREATE CONNECTION IF NOT EXISTS kafka_connection TO KAFKA (BROKER '{{ env_var('KAFKA_ADDR', 'localhost:9092') }}', SECURITY PROTOCOL PLAINTEXT)"
     )
 }}
  FROM {{ ref('test_materialized_view') }}
@@ -119,6 +126,20 @@ test_sink_cluster = """
 }}
  FROM {{ ref('test_materialized_view') }}
  INTO KAFKA CONNECTION kafka_connection (TOPIC 'test-sink')
+ FORMAT JSON
+ ENVELOPE DEBEZIUM
+"""
+
+test_sink_deploy = """
+{{ config(
+    materialized='sink',
+    schema='sinks_schema',
+    cluster='sinks_cluster',
+    pre_hook="CREATE CONNECTION IF NOT EXISTS kafka_connection TO KAFKA (BROKER '{{ env_var('KAFKA_ADDR', 'localhost:9092') }}', SECURITY PROTOCOL PLAINTEXT)"
+    )
+}}
+ FROM {{ ref('test_materialized_view_deploy') }}
+ INTO KAFKA CONNECTION kafka_connection (TOPIC 'testdrive-test-sink-1')
  FORMAT JSON
  ENVELOPE DEBEZIUM
 """