+++ misc/dbt-materialize/CHANGELOG.md
@@ -1,5 +1,19 @@
 # dbt-materialize Changelog
 
+## Unreleased
+
+* Add support for sink cutover to the blue/green deployment workflow [#27557](https://github.com/MaterializeInc/materialize/pull/27557).
+  Sinks **must** be created in a **dedicated schema and cluster**.
+
+* Add a `dry_run` argument to the `deploy_promote` macro, which allows
+  previewing the sequence of commands that will be run as part of the
+  environment promotion step of the blue/green deployment workflow.
+
+* Fix the `deploy_init` macro to correctly account for scheduled clusters.
+  Before, these clusters would be incorrectly recreated in the deployment
+  environment with the `SCHEDULE` option set to `manual` (instead of
+  `on-refresh`).
+
 ## 1.8.1 - 2024-06-08
 
 * Add support for overriding the `generate_cluster_name` macro to customize the